Add files via upload
fixed the window offset issues (FOR REAL FOR REAL) so varying window sizes should work correctly now. 
I also update some startup prompts so that they alow you to close them or quickly tab through the menu if you have already entered ionfo in the past, and happen to accidentally sya that you want to enter them again. Ill have a menu/gui for this in the future